Add guards for order presence
We detected an issue with some flash messages that are persisted in multiple requests
and that's causing to break pages where the @order variable is not present. This
makes sure it's present when rendering the snippets that make use of it.

Also normalizes the structure of the data used in the snippets.